+++ opentelemetry-propagator-gcp/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Fix propagator modifying context if failed to extract
+  ([#139](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/139))
+
 ## Version 1.0.0rc0
 
 Released 2021-04-22